Rename bfetch plugin (#37)
