--- src/plugins/bfetch/kibana.json
@@ -1,7 +1,0 @@
-{
-  "id": "bfetch",
-  "version": "kibana",
-  "server": true,
-  "ui": true,
-  "requiredBundles": ["kibanaUtils"]
-}
