+++ src/plugins/bfetch/opensearch_dashboards.json
@@ -1,0 +1,7 @@
+{
+  "id": "bfetch",
+  "version": "opensearchDashboards",
+  "server": true,
+  "ui": true,
+  "requiredBundles": ["opensearchDashboardsUtils"]
+}
