+++ src/plugins/bfetch/server/plugin.ts
@@ -23,11 +23,11 @@ import {
   CoreSetup,
   Plugin,
   Logger,
-  KibanaRequest,
+  OpenSearchDashboardsRequest,
   RouteMethod,
   RequestHandler,
 } from 'src/core/server';
-import { schema } from '@kbn/config-schema';
+import { schema } from '@osd/config-schema';
 import { Subject } from 'rxjs';
 import {
   StreamingResponseHandler,
@@ -41,10 +41,10 @@ import { StreamingRequestHandler } from './types';
 import { createNDJSONStream } from './streaming';
 
 // eslint-disable-next-line
-export interface BfetchServerSetupDependencies {}
+export interface BfetchServerSetupDependencies { }
 
 // eslint-disable-next-line
-export interface BfetchServerStartDependencies {}
+export interface BfetchServerStartDependencies { }
 
 export interface BatchProcessingRouteParams<BatchItemData, BatchItemResult> {
   onBatchItem: (data: BatchItemData) => Promise<BatchItemResult>;
@@ -54,11 +54,11 @@ export interface BatchProcessingRouteParams<BatchItemData, BatchItemResult> {
 export interface BfetchServerSetup {
   addBatchProcessingRoute: <BatchItemData extends object, BatchItemResult extends object>(
     path: string,
-    handler: (request: KibanaRequest) => BatchProcessingRouteParams<BatchItemData, BatchItemResult>
+    handler: (request: OpenSearchDashboardsRequest) => BatchProcessingRouteParams<BatchItemData, BatchItemResult>
   ) => void;
   addStreamingResponseRoute: <Payload, Response>(
     path: string,
-    params: (request: KibanaRequest) => StreamingResponseHandler<Payload, Response>
+    params: (request: OpenSearchDashboardsRequest) => StreamingResponseHandler<Payload, Response>
   ) => void;
   /**
    * Create a streaming request handler to be able to use an Observable to return chunked content to the client.
@@ -94,7 +94,7 @@ export interface BfetchServerSetup {
 }
 
 // eslint-disable-next-line
-export interface BfetchServerStart {}
+export interface BfetchServerStart { }
 
 const streamingHeaders = {
   'Content-Type': 'application/x-ndjson',
@@ -104,13 +104,13 @@ const streamingHeaders = {
 
 export class BfetchServerPlugin
   implements
-    Plugin<
-      BfetchServerSetup,
-      BfetchServerStart,
-      BfetchServerSetupDependencies,
-      BfetchServerStartDependencies
-    > {
-  constructor(private readonly initializerContext: PluginInitializerContext) {}
+  Plugin<
+  BfetchServerSetup,
+  BfetchServerStart,
+  BfetchServerSetupDependencies,
+  BfetchServerStartDependencies
+  > {
+  constructor(private readonly initializerContext: PluginInitializerContext) { }
 
   public setup(core: CoreSetup, plugins: BfetchServerSetupDependencies): BfetchServerSetup {
     const logger = this.initializerContext.logger.get();
@@ -130,7 +130,7 @@ export class BfetchServerPlugin
     return {};
   }
 
-  public stop() {}
+  public stop() { }
 
   private addStreamingResponseRoute = ({
     router,
@@ -181,32 +181,32 @@ export class BfetchServerPlugin
     E extends ErrorLike = ErrorLike
   >(
     path: string,
-    handler: (request: KibanaRequest) => BatchProcessingRouteParams<BatchItemData, BatchItemResult>
+    handler: (request: OpenSearchDashboardsRequest) => BatchProcessingRouteParams<BatchItemData, BatchItemResult>
   ) => {
-    addStreamingResponseRoute<
-      BatchRequestData<BatchItemData>,
-      BatchResponseItem<BatchItemResult, E>
-    >(path, (request) => {
-      const handlerInstance = handler(request);
-      return {
-        getResponseStream: ({ batch }) => {
-          const subject = new Subject<BatchResponseItem<BatchItemResult, E>>();
-          let cnt = batch.length;
-          batch.forEach(async (batchItem, id) => {
-            try {
-              const result = await handlerInstance.onBatchItem(batchItem);
-              subject.next({ id, result });
-            } catch (err) {
-              const error = normalizeError<E>(err);
-              subject.next({ id, error });
-            } finally {
-              cnt--;
-              if (!cnt) subject.complete();
-            }
-          });
-          return subject;
-        },
-      };
-    });
-  };
+      addStreamingResponseRoute<
+        BatchRequestData<BatchItemData>,
+        BatchResponseItem<BatchItemResult, E>
+      >(path, (request) => {
+        const handlerInstance = handler(request);
+        return {
+          getResponseStream: ({ batch }) => {
+            const subject = new Subject<BatchResponseItem<BatchItemResult, E>>();
+            let cnt = batch.length;
+            batch.forEach(async (batchItem, id) => {
+              try {
+                const result = await handlerInstance.onBatchItem(batchItem);
+                subject.next({ id, result });
+              } catch (err) {
+                const error = normalizeError<E>(err);
+                subject.next({ id, error });
+              } finally {
+                cnt--;
+                if (!cnt) subject.complete();
+              }
+            });
+            return subject;
+          },
+        };
+      });
+    };
 }
